Extract data from Table Component
